# public/swagger.json
{
	"components": {
...
		"schemas": {
...
			"ICreateResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IUpdateResponse": {
				"properties": {
					"id": {
						"anyOf": [
							{
								"type": "number",
								"format": "double"
							},
							{
								"type": "string"
							}
						]
					},
					"rows": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"rows"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IUser": {
				"properties": {
					"userId": {
						"type": "number",
						"format": "double"
					},
					"familyName": {
						"type": "string"
					},
					"givenName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"profileId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"userId",
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IIndexResponse_IUser_": {
				"properties": {
					"page": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"rows": {
						"items": {
							"$ref": "#/components/schemas/IUser"
						},
						"type": "array"
					}
				},
				"required": [
					"page",
					"limit",
					"total",
					"rows"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_IUser.Exclude_keyofIUser.userId__": {
				"properties": {
					"profileId": {
						"type": "number",
						"format": "double"
					},
					"familyName": {
						"type": "string"
					},
					"givenName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"email"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_IUser.userId_": {
				"$ref": "#/components/schemas/Pick_IUser.Exclude_keyofIUser.userId__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"IUserCreate": {
				"$ref": "#/components/schemas/Omit_IUser.userId_"
			},
			"Partial_IUserCreate_": {
				"properties": {
					"profileId": {
						"type": "number",
						"format": "double"
					},
					"familyName": {
						"type": "string"
					},
					"givenName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"IUserUpdate": {
				"$ref": "#/components/schemas/Partial_IUserCreate_"
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "authorization",
				"in": "header",
				"authorizationUrl": "http://swagger.io/api/oauth/dialog"
			}
		}
	},
	"info": {
		"title": "server",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/user": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IIndexResponse_IUser_"
								}
							}
						}
					}
				},
				"description": "Récupérer une page d'utilisateurs.",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "La page (zéro-index) à récupérer",
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Le nombre d'éléments à récupérer (max 50)",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ICreateResponse"
								}
							}
						}
					}
				},
				"description": "Créer un nouvel utilisateur",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserCreate"
							}
						}
					}
				}
			}
		},
		"/user/{userId}": {
			"get": {
				"operationId": "ReadUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUser"
								}
							}
						}
					}
				},
				"description": "Récupérer une utilisateur avec le ID passé dans le URL",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUpdateResponse"
								}
							}
						}
					}
				},
				"description": "Mettre à jour un utilisateur avec le ID passé dans le URL",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserUpdate"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUpdateResponse"
								}
							}
						}
					}
				},
				"description": "Supprimer un utilisateur",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/user/{userId}/file": {
			"post": {
				"requestBody": {
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"file": {
										"type": "string",
										"format": "binary"
									}
								}
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}